+++ plugin.video.3cat/README.md
@@ -1,0 +1,9 @@
+# plugin.video.3cat.cat
+Kodi addon 3Cat. 
+Entertainment, news, sports, documentaries, etc from catalan television www.3cat.cat/3cat/
+
+Complement per Kodi - XBMC.  
+Tota la programació de 3Cat.
+
+
+
