+++ plugin.video.3cat/changelog.txt
@@ -1,0 +1,2 @@
+﻿v1.0.0
+- First version
