+++ plugin.video.3cat/resources/language/resource.language.es_es/strings.po
@@ -1,0 +1,125 @@
+# Kodi Media Center language file
+# Addon Name: TV3cat
+# Addon id: plugin.video.3cat
+# Addon Provider: jqandreu
+msgid ""
+msgstr ""
+"Project-Id-Version: Kodi Addons\n"
+"Report-Msgid-Bugs-To: https://github.com/jqandreu/plugin.video.3cat/issues\n"
+"POT-Creation-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"Last-Translator: jqandreu\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: es_ES\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+msgctxt "#30001"
+msgid "Today Featured"
+msgstr "Hoy destacamos"
+
+msgctxt "#30002"
+msgid "Don't miss"
+msgstr "No te lo pierdas"
+
+msgctxt "#30003"
+msgid "Most viewed"
+msgstr "Lo más visto"
+
+msgctxt "#30004"
+msgid "Programs"
+msgstr "Programas"
+
+msgctxt "#30005"
+msgid "Series"
+msgstr "Series"
+
+msgctxt "#30006"
+msgid "News"
+msgstr "Informativos"
+
+msgctxt "#30007"
+msgid "Entertainment"
+msgstr "Entretenimiento"
+
+msgctxt "#30008"
+msgid "Sports"
+msgstr "Deportes"
+
+msgctxt "#30009"
+msgid "Documentaries"
+msgstr "Documentales"
+
+msgctxt "#30010"
+msgid "Divulgation"
+msgstr "Divulgación"
+
+msgctxt "#30011"
+msgid "Culture"
+msgstr "Cultura"
+
+msgctxt "#30012"
+msgid "Music"
+msgstr "Música"
+
+msgctxt "#30013"
+msgid "All A-Z"
+msgstr "Todos A-Z"
+
+msgctxt "#30014"
+msgid "Aired A-Z"
+msgstr "En emisión A-Z"
+
+msgctxt "#30015"
+msgid ">>Next Page"
+msgstr ">>Página siguiente"
+
+msgctxt "#30016"
+msgid ">>Previous Page"
+msgstr ">>Página anterior"
+
+msgctxt "#30017"
+msgid "Live channels"
+msgstr "Canales en directo"
+
+msgctxt "#30018"
+msgid "TV3"
+msgstr "TV3"
+
+msgctxt "#30019"
+msgid "Canal 324"
+msgstr "Canal 324"
+
+msgctxt "#30020"
+msgid "Canal 33-Super 3"
+msgstr "Canal 33-Super 3"
+
+msgctxt "#30021"
+msgid "Esport 3"
+msgstr "Esport 3"
+
+msgctxt "#30022"
+msgid "Search"
+msgstr "Buscar"
+
+msgctxt "#30023"
+msgid "Collections"
+msgstr "Colecciones"
+
+msgctxt "#30024"
+msgid "TV3 (International broadcast)"
+msgstr "TV3 (Emisión internacional)"
+
+msgctxt "#30025"
+msgid "Canal 324 (International broadcast)"
+msgstr "Canal 324 (Emisión internacional)"
+
+msgctxt "#30026"
+msgid "Canal 33-Super 3 (International broadcast)"
+msgstr "Canal 33-Super 3 (Emisión internacional)"
+
+msgctxt "#30027"
+msgid "Esport 3 (International broadcast)"
+msgstr "Esport 3 (Emisión internacional)"
