+++ plugin.video.3cat/resources/language/resource.language.nb_no/strings.po
@@ -1,0 +1,125 @@
+# Kodi Media Center language file
+# Addon Name: TV3cat
+# Addon id: plugin.video.3cat
+# Addon Provider: jqandreu
+msgid ""
+msgstr ""
+"Project-Id-Version: Kodi Addons\n"
+"Report-Msgid-Bugs-To: https://github.com/jqandreu/plugin.video.3cat/issues\n"
+"POT-Creation-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"Last-Translator: jrdfrncs\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: nb_NO\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+msgctxt "#30001"
+msgid "Today Featured"
+msgstr "Anbefalt i dag"
+
+msgctxt "#30002"
+msgid "Don't miss"
+msgstr "Ikke miss"
+
+msgctxt "#30003"
+msgid "Most viewed"
+msgstr "Mest sett"
+
+msgctxt "#30004"
+msgid "Programs"
+msgstr "Sendinger"
+
+msgctxt "#30005"
+msgid "Series"
+msgstr "Serier"
+
+msgctxt "#30006"
+msgid "News"
+msgstr "Nyheter"
+
+msgctxt "#30007"
+msgid "Entertainment"
+msgstr "Underholdning"
+
+msgctxt "#30008"
+msgid "Sports"
+msgstr "Idrett"
+
+msgctxt "#30009"
+msgid "Documentaries"
+msgstr "Dokumentaler"
+
+msgctxt "#30010"
+msgid "Divulgation"
+msgstr "Avsløring"
+
+msgctxt "#30011"
+msgid "Culture"
+msgstr "Kultur"
+
+msgctxt "#30012"
+msgid "Music"
+msgstr "Musikk"
+
+msgctxt "#30013"
+msgid "All A-Z"
+msgstr "Alt A-Å"
+
+msgctxt "#30014"
+msgid "Aired A-Z"
+msgstr "Nåværende A-Å"
+
+msgctxt "#30015"
+msgid ">>Next Page"
+msgstr ">>Neste side"
+
+msgctxt "#30016"
+msgid ">>Previous Page"
+msgstr ">>Forrige side"
+
+msgctxt "#30017"
+msgid "Live channels"
+msgstr "Direkte"
+
+msgctxt "#30018"
+msgid "TV3"
+msgstr "TV3"
+
+msgctxt "#30019"
+msgid "Canal 324"
+msgstr "Canal 324"
+
+msgctxt "#30020"
+msgid "Canal 33-Super 3"
+msgstr "Canal 33-Super 3"
+
+msgctxt "#30021"
+msgid "Esport 3"
+msgstr "Esport 3"
+
+msgctxt "#30022"
+msgid "Search"
+msgstr "Søk"
+
+msgctxt "#30023"
+msgid "Collections"
+msgstr "Samlinger"
+
+msgctxt "#30024"
+msgid "TV3 (International broadcast)"
+msgstr "TV3 (Utland)"
+
+msgctxt "#30025"
+msgid "Canal 324 (International broadcast)"
+msgstr "Canal 324 (Utland)"
+
+msgctxt "#30026"
+msgid "Canal 33-Super 3 (International broadcast)"
+msgstr "Canal 33-Super 3 (Utland)"
+
+msgctxt "#30027"
+msgid "Esport 3 (International broadcast)"
+msgstr "Esport 3 (Utland)"
