+++ plugin.video.3cat/resources/lib/utils/Urls.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+
+url_base = 'https://www.3cat.cat/3cat/'
+url_alacarta = 'http://www.ccma.cat/tv3/alacarta/'
+url_coleccions = 'https://www.3cat.cat/3cat/tot-cataleg/tot/'
+url_mesvist = 'http://www.ccma.cat/tv3/alacarta/mes-vist/'
+url_datavideos = 'http://dinamics.ccma.cat/pvideo/media.jsp?media=video&version=0s&idint='
+url_programes_emisio = 'http://www.ccma.cat/tv3/alacarta/programes'
+url_programes_tots = 'http://www.ccma.cat/tv3/alacarta/programes-tots/'
+urlZonaZaping = 'http://www.ccma.cat/tv3/alacarta/zona-zaping/'
+urlApm = 'http://www.ccma.cat/tv3/alacarta/apm/'
+url_directe_tv3 = 'https://directes-tv-es.ccma.cat/es/ngrp:tvc_web/playlist.m3u8'
+url_directe_324 = 'https://directes-tv-int.ccma.cat/int/ngrp:324_web/playlist.m3u8'
+url_directe_c33s3 = 'https://directes-tv-es.ccma.cat/es/ngrp:c33_web/playlist.m3u8'
+url_directe_esport3 = 'https://directes-tv-es.ccma.cat/es/ngrp:es3_web/playlist.m3u8'
+#Feeds per a fer streaming des de l'estranger
+url_directe_tv3_int = 'https://directes-tv-int.ccma.cat/int/ngrp:tvi_web/playlist.m3u8'
+url_directe_324_int = 'https://directes-tv-int.ccma.cat/int/ngrp:324_web/playlist.m3u8'
+url_directe_c33s3_int = 'https://directes-tv-int.ccma.cat/int/ngrp:c33_web/playlist.m3u8'
+url_directe_esport3_int = 'https://directes-tv-int.ccma.cat/int/ngrp:es3_web/playlist.m3u8'
+
+url_arafem ='http://dinamics.ccma.cat/wsarafem/arafem/tv'
+
