--- plugin.video.tv3.cat/resources/language/resource.language.ca_es/strings.po
@@ -1,125 +1,0 @@
-# Kodi Media Center language file
-# Addon Name: TV3cat
-# Addon id: plugin.video.tv3.cat
-# Addon Provider: jqandreu
-msgid ""
-msgstr ""
-"Project-Id-Version: Kodi Addons\n"
-"Report-Msgid-Bugs-To: https://github.com/jqandreu/plugin.video.tv3.cat/issues\n"
-"POT-Creation-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: jqandreu\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Language: es_CA\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-
-msgctxt "#30001"
-msgid "Today Featured"
-msgstr "Avui destaquem"
-
-msgctxt "#30002"
-msgid "Don't miss"
-msgstr "No t'ho perdis"
-
-msgctxt "#30003"
-msgid "Most viewed"
-msgstr "El més vist"
-
-msgctxt "#30004"
-msgid "Programs"
-msgstr "Programes"
-
-msgctxt "#30005"
-msgid "Series"
-msgstr "Sèries"
-
-msgctxt "#30006"
-msgid "News"
-msgstr "Informatius"
-
-msgctxt "#30007"
-msgid "Entertainment"
-msgstr "Entreteniment"
-
-msgctxt "#30008"
-msgid "Sports"
-msgstr "Esports"
-
-msgctxt "#30009"
-msgid "Documentaries"
-msgstr "Documentals"
-
-msgctxt "#30010"
-msgid "Divulgation"
-msgstr "Divulgació"
-
-msgctxt "#30011"
-msgid "Culture"
-msgstr "Cultura"
-
-msgctxt "#30012"
-msgid "Music"
-msgstr "Música"
-
-msgctxt "#30013"
-msgid "All A-Z"
-msgstr "Tots A-Z"
-
-msgctxt "#30014"
-msgid "Aired A-Z"
-msgstr "En emissió A-Z"
-
-msgctxt "#30015"
-msgid ">>Next Page"
-msgstr ">>Pàgina següent"
-
-msgctxt "#30016"
-msgid ">>Previous Page"
-msgstr ">>Pàgina anterior"
-
-msgctxt "#30017"
-msgid "Live channels"
-msgstr "Canals en directe"
-
-msgctxt "#30018"
-msgid "TV3"
-msgstr "TV3"
-
-msgctxt "#30019"
-msgid "Canal 324"
-msgstr "Canal 324"
-
-msgctxt "#30020"
-msgid "Canal 33-Super 3"
-msgstr "Canal 33-Super 3"
-
-msgctxt "#30021"
-msgid "Esport 3"
-msgstr "Esport 3"
-
-msgctxt "#30022"
-msgid "Search"
-msgstr "Cercar"
-
-msgctxt "#30023"
-msgid "Collections"
-msgstr "Col·leccions"
-
-msgctxt "#30024"
-msgid "TV3 (International broadcast)"
-msgstr "TV3 (Emissió internacional)"
-
-msgctxt "#30025"
-msgid "Canal 324 (International broadcast)"
-msgstr "Canal 324 (Emissió internacional)"
-
-msgctxt "#30026"
-msgid "Canal 33-Super 3 (International broadcast)"
-msgstr "Canal 33-Super 3 (Emissió internacional)"
-
-msgctxt "#30027"
-msgid "Esport 3 (International broadcast)"
-msgstr "Esport 3 (Emissió internacional)"
