--- plugin.video.tv3.cat/resources/lib/video/FolderVideo.py
@@ -1,16 +1,0 @@
-from builtins import object
-class FolderVideo(object):
-    def __init__(self, name, url, mode, iconImage, thumbnaiImage):
-        """
-
-        :rtype: object
-        """
-
-        self.name = name
-        self.url = url
-        self.mode = mode
-        self.iconImage = iconImage
-        self.thumbnailImage = thumbnaiImage
-
-
-
